+++ src/lib.js
@@ -284,3 +284,14 @@ exports.keys = function(obj) {
         return keys;
     }
 };
+
+exports.inOperator = function (key, arrOrObj) {
+    if (exports.isArray(arrOrObj)) {
+        return exports.indexOf(arrOrObj, key) !== -1;
+    } else if (exports.isObject(arrOrObj)) {
+        return key in arrOrObj;
+    } else {
+        throw new Error('Cannot use "in" operator to search for "'
+            + key + '" in unexpected types.');
+    }
+};